fix(auth): use float comparison for password_changed_at epoch to avoid sub-second truncation
token_iat (JWT iat) is an integer second, while password_changed_at has
sub-second precision from datetime.timestamp(). int() truncation meant
tokens created in the same second as the password change were not rejected.

+++ backend/app/api/deps.py
@@ -442,7 +442,7 @@ async def get_current_active_user(
     # (real password-change events). Login rehash is NOT a password change.
     pwd_epoch = user.get("password_changed_at") or 0
     token_iat = payload.get("iat", 0) if isinstance(payload, dict) else 0
-    if pwd_epoch > 0 and token_iat < int(pwd_epoch):
+    if pwd_epoch > 0 and float(token_iat) < pwd_epoch:
         invalidate_active_user_cache(user_id)
         raise HTTPException(
             status_code=401,
